# config/openapi/service/openapi.yaml
openapi: 3.0.3
info:
  title: HPC Job Observability Service API
  description: |
    A job-level monitoring and observability service for HPC environments.
    This API provides access to job metadata, states, and resource usage metrics.
  version: 1.0.0
  contact:
    name: HPC Observability Team
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:8080/v1
    description: Local development server
...
paths:
...
  /jobs:
    get:
      tags:
        - jobs
      summary: List all jobs
      description: Returns a list of all known jobs with optional filtering
      operationId: listJobs
      parameters:
        - name: state
          in: query
          description: Filter by job state
          required: false
          schema:
            $ref: '#/components/schemas/JobState'
        - name: user
          in: query
          description: Filter by user
          required: false
          schema:
            type: string
            example: user123
        - name: node
          in: query
          description: Filter by node name
          required: false
          schema:
            type: string
            example: node-01
        - name: limit
          in: query
          description: Maximum number of jobs to return
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: offset
          in: query
          description: Number of jobs to skip for pagination
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobListResponse'
              example:
                jobs:
                  - id: job-001
                    user: alice
                    nodes:
                      - node-01
                      - node-02
                    state: running
                    start_time: '2026-01-20T08:00:00Z'
                    cpu_usage: 85.5
                    memory_usage_mb: 4096
                    gpu_usage: 45.0
                  - id: job-002
                    user: bob
                    nodes:
                      - node-03
                    state: completed
                    start_time: '2026-01-19T14:00:00Z'
                    end_time: '2026-01-19T18:30:00Z'
                    runtime_seconds: 16200
                    cpu_usage: 0
                    memory_usage_mb: 0
                total: 2
                limit: 100
                offset: 0
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - jobs
      summary: Create a new job
      description: Register a new job in the system
      operationId: createJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
            example:
              id: job-003
              user: charlie
              nodes:
                - node-04
                - node-05
      parameters:
        - $ref: '#/components/parameters/ChangedByHeader'
        - $ref: '#/components/parameters/SourceHeader'
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '201':
          description: Job created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Job with this ID already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /jobs/{jobId}:
    get:
      tags:
        - jobs
      summary: Get job details
      description: Returns metadata and current state for a specific job
      operationId: getJob
      parameters:
        - $ref: '#/components/parameters/JobId'
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
              example:
                id: job-001
                user: alice
                nodes:
                  - node-01
                  - node-02
                state: running
                start_time: '2026-01-20T08:00:00Z'
                cpu_usage: 85.5
                memory_usage_mb: 4096
                gpu_usage: 45.0
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    patch:
      tags:
        - jobs
      summary: Update job
      description: Update job state or metadata
      operationId: updateJob
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/ChangedByHeader'
        - $ref: '#/components/parameters/SourceHeader'
        - $ref: '#/components/parameters/CorrelationIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateJobRequest'
            example:
              state: completed
      responses:
        '200':
          description: Job updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - jobs
      summary: Delete job
      description: Remove a job from the system
      operationId: deleteJob
      parameters:
        - $ref: '#/components/parameters/JobId'
        - $ref: '#/components/parameters/ChangedByHeader'
        - $ref: '#/components/parameters/SourceHeader'
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '204':
          description: Job deleted successfully
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /events/job-started:
    post:
      tags:
        - events
      summary: Job started event
      description: |
        Receives a job started event from Slurm prolog scripts.
        This event registers a new job with its cgroup path and GPU allocation.
        The endpoint is idempotent - duplicate events are handled safely.
      operationId: jobStartedEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobStartedEvent'
            example:
              job_id: "12345"
              user: alice
              node_list:
                - node-01
                - node-02
              cpu_allocation: 16
              gpu_allocation: 2
              gpu_vendor: nvidia
              gpu_devices:
                - GPU-abc123
                - GPU-def456
              cgroup_path: /sys/fs/cgroup/system.slice/slurmstepd.scope/job_12345
              timestamp: '2026-01-22T10:00:00Z'
      responses:
        '200':
          description: Job started event processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobEventResponse'
        '400':
          description: Invalid event payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  parameters:
...
    ChangedByHeader:
      name: X-Changed-By
      in: header
      description: Identifier of the actor/system performing the change
      required: true
      schema:
        type: string
        minLength: 1
    SourceHeader:
      name: X-Source
      in: header
      description: Source system initiating the change (api, syncer, scheduler)
      required: true
      schema:
        type: string
        minLength: 1
    CorrelationIdHeader:
      name: X-Correlation-Id
      in: header
      description: Correlation ID for tracing multi-step updates
      required: false
      schema:
        type: string
  schemas:
...
    JobState:
      type: string
      enum:
        - pending
        - running
        - completed
        - failed
        - cancelled
      description: Current state of the job
...
    CreateJobRequest:
      type: object
      required:
        - id
        - user
        - nodes
      properties:
        id:
          type: string
          minLength: 1
          maxLength: 255
          description: Unique job identifier
        user:
          type: string
          minLength: 1
          maxLength: 255
          description: User who submitted the job
        nodes:
          type: array
          items:
            type: string
          minItems: 1
          description: List of node names to allocate
        scheduler:
          $ref: '#/components/schemas/SchedulerInfo'
...
    JobListResponse:
      type: object
      required:
        - jobs
        - total
        - limit
        - offset
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/Job'
        total:
          type: integer
          description: Total number of jobs matching the filter
        limit:
          type: integer
          description: Maximum number of jobs returned
        offset:
          type: integer
          description: Number of jobs skipped
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: not_found
        message:
          type: string
          description: Human-readable error message
          example: Job with ID 'job-999' not found
        details:
          type: object
          additionalProperties: true
          description: Additional error details
...